+++ breeze-core/src/main/kotlin/domain/Configurable.kt
@@ -8,7 +8,7 @@ package com.windea.breezeframework.core.domain
  */
 interface Configurable<T> {
 	/**
-	 * 基于特定的配置对象，配置当前对象。
+	 * 基于特定的配置对象，配置当前对象。一般来说，需要在使用前配置完毕。
 	 */
 	fun configure(block:T.()->Unit)
 }