+++ breeze-serialization/src/main/kotlin/CsvSerializer.kt
@@ -14,6 +14,13 @@ interface CsvSerializer : Serializer{
 	//region Csv Serializers
 	/**
 	 * 默认的Csv序列化器。
+	 *
+	 * 可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
+	 */
+	companion object Default:CsvSerializer by defaultCsvSerializer
+
+	/**
+	 * 框架本身实现的Csv序列化器。
 	 */
 	object BreezeCsvSerializer:CsvSerializer{
 		override fun <T : Any> serialize(value: T): String {