+++ breeze-serialization/src/main/kotlin/SerializationExtensions.kt
@@ -1,6 +1,8 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
+@file:JvmName("SerializationExtensions")
+
 package com.windea.breezeframework.serialization
 
 import com.windea.breezeframework.core.annotations.*
@@ -10,20 +12,20 @@ import java.lang.reflect.*
 /**
  * 根据指定的数据类型，序列化当前对象。
  *
- * 可以由第三方库委托实现，基于classpath，推断具体实现。默认使用框架本身的简单实现，可以自定义。
+ * 这个方法使用的序列化器可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
  *
  * @see DataType
  * @see Serializer
  */
 @BreezeComponentExtension
-fun <T:Any> T.serializeBy(dataType:DataType):String{
+fun <T:Any> T.serializeBy(dataType: DataType):String{
 	return dataType.serialize(this)
 }
 
 /**
  * 根据指定的数据类型，反序列化当前文本。
  *
- * 可以由第三方库委托实现，基于classpath，推断具体实现。默认使用框架本身的简单实现，可以自定义。
+ * 这个方法使用的序列化器可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
  *
  * @see DataType
  * @see Serializer
@@ -36,7 +38,7 @@ inline fun <reified T:Any> String.deserializeBy(dataType:DataType):T{
 /**
  * 根据指定的数据类型，反序列化当前文本。
  *
- * 可以由第三方库委托实现，基于classpath，推断具体实现。默认使用框架本身的简单实现，可以自定义。
+ * 这个方法使用的序列化器可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
  *
  * @see DataType
  * @see Serializer
@@ -49,7 +51,7 @@ fun <T:Any> String.deserializeBy(dataType: DataType,type:Class<T>):T{
 /**
  * 根据指定的数据类型，反序列化当前文本。
  *
- * 可以由第三方库委托实现，基于classpath，推断具体实现。默认使用框架本身的简单实现，可以自定义。
+ * 这个方法使用的序列化器可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
  *
  * @see DataType
  * @see Serializer