+++ breeze-serialization/src/main/kotlin/XmlSerializer.kt
@@ -16,6 +16,13 @@ interface XmlSerializer : Serializer {
 
 	//region Xml Serializers
 	/**
+	 * 默认的Xml序列化器。
+	 *
+	 * 可以由第三方库委托实现，基于classpath进行推断，或者使用框架本身实现的序列化器。
+	 */
+	companion object Default:XmlSerializer by defaultXmlSerializer
+
+	/**
 	 * 由Jackson实现的Xml序列化器。
 	 *
 	 * @see com.fasterxml.jackson.dataformat.xml.XmlMapper
@@ -45,7 +52,7 @@ interface XmlSerializer : Serializer {
 	}
 
 	/**
-	 * 默认的Xml序列化器。
+	 * 框架本身实现的Xml序列化器。
 	 */
 	object BreezeXmlSerializer:XmlSerializer{
 		override fun <T : Any> serialize(value: T): String {